requirements.txt
markitdown[pdf]>=0.0.1a3
PyYAML>=6.0
python-frontmatter>=1.1.0
openai>=1.40.0
